+++ src/Core/OakVR/Updateable.h
@@ -7,6 +7,6 @@ namespace oakvr
 	public:
 		Updateable() {}
 		virtual ~Updateable() {}
-		virtual void Update(double dt) = 0; /* override in derived class to add custom behavior */
+		virtual void Update(float dt) = 0; /* override in derived class to add custom behavior */
 	};
 }
